Switch to unified, bounds checked, `Reg` to name registers
This change will pay off when we amortize the decoding.

The unification is debatable; it's advantageous in saving code in
places, but it costs one extra bit.  For the SW implementation this is
a non-issue and it's would be trivial (yet tedious) to split this into
`Reg` and `FReg`.  In future perhaps.